add: UIImage extension
